+++ include/cppcsv/simplecsv.hpp
@@ -15,7 +15,7 @@ public:
   virtual ~csv_builder() {}
 
   virtual void begin_row() = 0;
-  virtual void cell( const char *buf, unsigned int len ) = 0;  // buf can be NULL
+  virtual void cell( const char *buf, size_t len ) = 0;  // buf can be NULL
   virtual void end_row() = 0;
 };
 
@@ -33,14 +33,14 @@ public:
 private:
   friend class Table;
   friend class Row;
-  Value(const Row &parent, unsigned int cidx,const std::string &value);
+  Value(const Row &parent, size_t cidx,const std::string &value);
 
   Value(const Value&); // = delete
   Value &operator=(const Value &);
 private:
   // parent, cidx  are currently unused
   const Row &parent;
-  unsigned int cidx;
+  size_t cidx;
 
   std::string value;
 };
@@ -52,25 +52,25 @@ public:
   ~Row();
 
   const Value &operator[](const char *key) const;
-  const Value &operator[](unsigned int cidx) const;
+  const Value &operator[](size_t cidx) const;
 
-  unsigned int size() const;
+  size_t size() const;
 
   void dump() const;
 public:
   // special case: &value==&del
-  void set(unsigned int cidx, const std::string &value);  // non-const !
+  void set(size_t cidx, const std::string &value);  // non-const !
 
   static const std::string del; // sentinel
 private:
   friend class Table;
-  Row(Table &parent, unsigned int ridx);
+  Row(Table &parent, size_t ridx);
   void write(csv_builder &out) const;
 private:
   Table &parent;
-  unsigned int ridx;
+  size_t ridx;
 
-  std::map<unsigned int,Value *> columns;  // TODO: not by *? (need move?)
+  std::map<size_t,Value *> columns;  // TODO: not by *? (need move?)
 };
 
 class Table {
@@ -80,16 +80,16 @@ public:
   Table() {} // = default
   ~Table();
 
-  const Row &operator[](unsigned int ridx) const;
-  unsigned int size() const;
+  const Row &operator[](size_t ridx) const;
+  size_t size() const;
 
   void dump() const;
 public:
   class IBuild {
   public:
     static Row &newRow(Table &csv);
-    static Row &insertRow(Table &csv, unsigned int at_ridx); // 0<at_ridx<=size();   // replaces! (TODO? real insert?)
-    static void deleteRow(Table &csv, unsigned int ridx);  // (moves elements!)
+    static Row &insertRow(Table &csv, size_t at_ridx); // 0<at_ridx<=size();   // replaces! (TODO? real insert?)
+    static void deleteRow(Table &csv, size_t ridx);  // (moves elements!)
 
     static void setHeader(Table &csv,const std::vector<std::string>& names);
   };
@@ -103,7 +103,7 @@ private:
   static const Row empty_row;      // sentinel
   static const Value empty_value;  // sentinel
 private:
-  std::multimap<std::string, unsigned int,lt_nocase_str> rev_column;
+  std::multimap<std::string, size_t,lt_nocase_str> rev_column;
   std::vector<const std::string *> columnnames;
   std::vector<Row *> rows;  // CPP11:  std::unique_ptr!   OR  by-value(Row moveable)?
 };
@@ -112,12 +112,12 @@ class builder : public per_cell_tag {
 public:
   builder(Table &result,bool first_is_header=false);
   void begin_row();
-  void cell(const char *buf, unsigned int len);
+  void cell(const char *buf, size_t len);
   void end_row();
 private:
   Table &result;
   Row *row;
-  unsigned int cidx;
+  size_t cidx;
   bool as_header;
   std::vector<std::string> header;
 };
